Fixed calculation of startTs in getReservations()
When having more than seven intervals per day the calculation of
startOfDay didn't work since it used the variable k used for
calculating the amount of bytes needed to be read from the server,
which was the same as the number of intervals up until 7. After that,
we get an off-by-one.

Let's use nintervals instead, which already holds the correct value.